% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{call_airtable}
\alias{call_airtable}
\title{Airtable Caller}
\usage{
call_airtable(base, table, token, query)
}
\arguments{
\item{base}{Airtable base ID e.g. \code{"app5gby0DG3pF9KbB"}}

\item{table}{Airtable tablename e.g. \code{"tablename"}}

\item{token}{Airtable API token}

\item{query}{A list of query parameters e.g. \code{list("filterByFormula"="(appsboard=1)")}}
}
\value{
JSON data if request was successful.
}
\description{
Fetches data from a single call to Airtable API into JSON.
}
\examples{
json = call_airtable(AT_BASE, AT_TABLE, AT_TOKEN, query=list("filterByFormula"="(appsboard=1)"))
}
\seealso{
\code{\link[=fetch_airtable]{fetch_airtable()}} which is built on top of this function.
}
