% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{download_feather}
\alias{download_feather}
\title{Feather from URL Fetcher}
\usage{
download_feather(url)
}
\arguments{
\item{url}{A direct URL of Feather file}
}
\value{
A dataframe with Feather file's data.
}
\description{
Fetches Feather file from URL into dataframe.
}
\examples{
dataframe = download_feather("http://example.com/data/table.feather")
}
