% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{fetch_airtable}
\alias{fetch_airtable}
\title{Airtable Parser}
\usage{
fetch_airtable(..., query = list())
}
\arguments{
\item{query}{A list of query parameters e.g. \code{list("filterByFormula"="(appsboard=1)")}}

\item{base}{Airtable base ID e.g. \code{"app5gby0DG3pF9KbB"}}

\item{table}{Airtable tablename e.g. \code{"tablename"}}

\item{token}{Airtable API token}
}
\value{
A dataframe with Airtable data.
}
\description{
Fetches data from Airtable API into dataframe.
}
\examples{
at = fetch_airtable(AT_BASE, AT_TABLE, AT_TOKEN, query=list("filterByFormula"="(appsboard=1)"))
}
\seealso{
\code{\link[=call_airtable]{call_airtable()}} which this function wraps.
}
